added cache

+++ src/main/java/yandex/eventservice/api/EventController.java
@@ -24,7 +24,7 @@ public class EventController extends RestController {
     }
 
     @GetMapping
-    public BigInteger getEvents(@RequestParam(required = false, defaultValue = "0") Integer minutes,
+    public Long getEvents(@RequestParam(required = false, defaultValue = "0") Integer minutes,
                                 @RequestParam(required = false, defaultValue = "0") Integer hours) {
         return eventService.countLast(hours * 60 + minutes);
     }
